+++ src/main/java/de/adorsys/keycloak/config/service/UserImportService.java
@@ -140,7 +140,10 @@ public class UserImportService {
         private void updateUser(UserRepresentation existingUser) {
             UserRepresentation patchedUser = CloneUtil
                     .deepPatch(existingUser, userToImport, IGNORED_PROPERTIES_FOR_UPDATE);
-            if (userToImport.getAttributes() != null) {
+
+            if (importConfigProperties.isSkipAttributesForFederatedUser() && patchedUser.getFederationLink() != null) {
+                patchedUser.setAttributes(null);
+            } else if (existingUser.getAttributes() != null && userToImport.getAttributes() != null) {
                 patchedUser.setAttributes(userToImport.getAttributes());
             }